requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
binance-connector>=3.5.1
numpy>=1.26.2
pandas>=2.0.0
scipy>=1.11.4
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.1
websockets>=12.0
scikit-learn>=1.3.2
